requirements-old.txt
streamlit-oauth==0.1.8
httpx-oauth
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.2
annoy==1.17.2
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.6
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
authenticator==1.1.3
autopep8==2.0.1
axios==0.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.2
cloudpickle==2.2.1
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
dnspython==2.3.0
duckdb==0.8.0
exceptiongroup==1.1.1
executing==1.2.0
extra-streamlit-components==0.1.56
fastapi==0.97.0
fastjsonschema==2.17.1
filelock==3.12.1
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
fonttools==4.42.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future>=0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
graphviz==0.20.1
grpcio==1.56.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.9.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
httpx-oauth==0.13.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.1
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso8601==2.0.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
kiwisolver==1.4.5
langchain==0.0.195
langchainplus-sdk==0.0.8
lazy-object-proxy==1.9.0
libclang==16.0.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
macholib>=1.15.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
neat-python==0.92
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.9.0
penngrader-client==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==2023.6.12
platformdirs==3.5.3
posthog==3.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pydeck==0.8.1b0
pygame==2.4.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.5
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.9.1
PyPDF2==3.0.1
PyPDF4==1.27.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six>=1.15.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
streamlit==1.28.0
streamlit-aggrid==0.3.4
streamlit-authenticator==0.2.2
streamlit-chat==0.1.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchsummary==1.5.1
torchvision==0.15.1
torchviz==0.0.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
trio==0.22.0
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0